Order of subroutines swapped
Routine affected - gcmc_control.f90

Get_Fugacity_Info and Get_Zig_By_Omega routines are called after
Get_Move_Probabilities is called. It was necessary to change the order
so that Get_Move_Probabilities properly assigns the int_sorbate_type to
0 indicating that the species can be swapped. This is required so that
the values of activity and zig/omega are properly set to the user input
values. Previously, they would be assigned a value of 0.0 for these
variables.