Add faujasite example
The PDB file in this commit was obtained by external collaborator who was
getting errors when executing mcfgen.py. The file contains
a CONECT section that was previously used to build molecule
connectivity information of the zeolite framework. Since
we do not support rigid frameworks, this information is
irrelevant and was actually producing an incorrect MCF.

Additionally, the original PDB file was not consistent with the
PDB format specification. In particular, the element symbol
was not in the 77-78 columns, which now is.