Add trajectory reader and enhance Widom insertion support. (#106)
This PR allows Cassandra to utilize trajectories generated from LAMMPS as an input to a Widom insertion calculation, which allows the user to compute Henry's constants. The LAMMPS trajectories are first converted to XYZ and H files with a Python script, after which Cassandra reads these in and performs Widom insertions in parallel along the trajectory. A Cell-list based overlap detection method is used to make these insertions more efficient.  New output files whose extensions is wprp are generated with the Widom variable results.

This PR also includes various changes such as using 64 -bit integers to be able to run for longer simulations and accomodate larger Widom input parameters. 

Documentation and example input files are included.

Co-authored-by: Eliseo Marin-Rimoldi <emarinri@nd.edu>