Add apparently working bitcell overlap detection and CBMC cell neighbor list.
Members of gathered overlap cells and cell neighbor lists are now filtered by proximity.
CBMC cell list option would now be more appropriately called a cell neighbor list method, since
the possible neighbors for a cell are now gathered and filtered by proximity.  CBMC cells are now
the same size as overlap cells; the gathering algorithm just searches more cells to capture all possible
neighbors.  Trial insertion of first fragment in CBMC are now greatly vectorized.  CBMC dihedral trials are not yet,
but applying vectorization and bitcell overlap detection to dihedral trials should be fairly straightforward.
Dimension padding currently assumes vector size no greater than 256 bits (the size of AVX2 vector registers), and if
we want Cassandra to support AVX-512, changes need to be made to accommodate that since it would violate the alignment
assumptions made in some ifort compiler directives.  While intermolecular CBMC energy estimation is vectorized when used
with CBMC cell neighbor lists, it can apparently sometimes still be slightly slower than directly computing the energy,
most likely due to slower memory access for the very large, precomputed energy table.  I still left it as an option though because
for more expensive force fields, it may be faster.  Some cheap WRITE statements used for debugging are still present in the code
and should probably be removed to avoid excessive verbosity, especially to STDOUT.
Repeating an old simulation (from before this commit) using the same seeds and simulation options will not give identical
results even with a single thread due to the way CBMC insertion trial positions are calculated from the random numbers
differing from how it used to be done; for example, using rranf() - 0.5 instead of 0.5 - rranf() as fractional COM coordinate.
Restricted insertion trial coordinates are now generated within the inner volume the first time, rather than
being generated anywhere in the box and re-generating them within the inner volume them if they're outside the inner volume, as
was done previously, and this process is now vectorized.  Widom insertions will no longer be restricted ever, even if the inserted
species is designated with restricted GCMC insertions. It's likely this was never a problem for anyone, but this fix should make sure
it won't be a problem in the future.  If restricted Widom insertions are ever allowed in the future, additional changes will need
to be made for it to be done properly.