Add support for RB torsion dihedrals and convert other dihedral types to RB form where possible.
All OPLS dihedrals are internally converted to RB torsions now because RB torsions are much faster to compute.
CHARMM style dihedrals are converted to RB torsions when it is possible to do so (I don't think I've seen one that
isn't possible to convert to RB but they might exist), and they are left as CHARMM style if it isn't possible to convert to RB format.
All dihedrals formatted as RB torsions (whether explicitly input or internally converted) that are stacked on the same or reverse
4-atom sequence as each other are collapsed into a single RB torsion by adding together the coefficients of the stacked RB torsions.
RB torsions are implemented in the protein convention (based on phi) in Cassandra, like the other dihedral types are.
This differs from how they are implemented in GROMACS, which uses the polymer convention (based on psi, which is phi - pi).
To convert from one convention to the other (either direction), simply flip the sign of the coefficients of the
even-powered terms of the series.

I also commented out the code that reads parameters for AMBER-style dihedrals because Cassandra has no
code to compute the energies of AMBER-style dihedrals and they aren't converted to another style either.
Dihedral styles are now allowed to be specified in all-caps or all lowercase in the mcf files, to make things more user-friendly.
I also renamed get_internal_coords.f90 to internal_coordinate_routines.f90 and made Internal_Coordinate_Routines a module, since
the file previously just contained a collection of subroutines, one of which is named Get_Internal_Coords, not encompassed by a module.