Refactor fragment_placement, enabling bitcell overlap detection and more vectorization.
Also designate several procedures as ELEMENTAL for ease of use and optimization.
Add optional argument l_skip_dihed_vec to Compute_Molecule_Dihedral_Energy that specifies
which dihedrals to skip computing energy for.
Allow different species to use different CBMC kappa values.  Add way to specify minimum
ideal_bitcell_length, which overrides the ideal_bitcell_length computed by the default method
if it is greater than the computed ideal_bitcell_length, but not if it is smaller, since
the computed value is the minimum value required for the algorithm to work properly.  The
user-defined minimum is an option because it can be beneficial to lower the resolution of the
bitcell grid so it occupies less memory and allows faster memory access and probably has a
lower cache miss rate. This will result in the bitcell overlap method catching fewer overlaps
(which will be caught by the cell list overlap detection instead if they are overlaps), but
allowing the bitcells to be checked faster can be worth it (tested with min_ideal_bitcell_length = 0.2).
New function Excess_Molecule_Intrafragment_Energy was added, and optional excess_flag_o and/or minimg_flag_o arguments
were added to a few subroutines to cause them to instead compute the "excess" energy (energy minus what it would be if
computed with the minimum image sum style as during fragment library generation) and minimum image energy
(essentially forces the subroutine to act as if the sum style is minimum image, even if it isn't) so you only get
the intramolecular parts you need for Widom insertions.
Interfragment intramolecular energy is now optionally output by Build_Molecule as E_interfrag, though the logic in
Fragment_Placement causes it to only do so during Widom insertions.  That should probably be changed if/when the new
intramolecular energy accounting done in Widom insertions is applied to other CBMC moves.  Widom insertions now include
no intramolecular energy except for what is computed by Excess_Molecule_Intrafragment_Energy and the interfragment
intramolecular energy, since any remaining parts would have been used to generate the fragment libraries and would
have to be subtracted back out if included, which is inefficient.  This method should be more robust than what was
previously done, and should probably be applied to other CBMC moves as well.
Use of undamped shifted force method for coulombic interactions in CBMC trial energies is only partially implemented.