Optimize Ewald summation code.
Improve vectorization and multithreading and improve mathematical formulation of Ewald summation.
Also overhaul data structures for Ewald data and molecule pair energy and replace
large array copying with memory allocation transfers or remove them entirely (if they're unnecessary).
The arbitrary limit on the number of kspace vectors was removed and replaced with a much larger limit
based on implementation limitations that are unlikely to be met for sane systems.
If the new limits ever become too low for a sane application, the limits may be increased by updating
the integer components of a kspace vector to be encoded in a 64-bit integer instead of a 32-bit integer.
For triclinic and non-cubic, orthogonal boxes, the range within which to check kspace vectors is automatically
computed based on the face distances of a box in reciprocal space for which the cell matrix is
the transpose of the inverse of the cell matrix for the real box.
Previously, the range to check in reciprocal space was hardcoded for triclinic and non-cubic, orthogonal boxes.